requirements.txt
filemagic
sh
hachoir3
patool
